(feat) Create function for getting user's contributable caches
In this current version, we are passing in the id of the
user, and using a callback.

This function is automatically called when the user logs in.
In the future, we will probably want to use promises, and somehow
just use the current user (instead of passing in the user's id).