(feat) Create initial skeleton for UserFriendsService
Initially, this service will be used to get all of the users in our
app, and allow the user to search through that list.

In the future, we will want to change this to only be their
Facebook friends who also have an account.

(feat) Create getAllUsers() to grab all user objects from Firebase

This function is currently doing a somewhat stupid way, in the sense
it is grabbing all of the data, which will then be processed on the
client-side.

For MVP, this is probably fine, but most likely we will want to
use another Firebase collection that will make these operations
easier.

(feat) Create helper _getDisplayNames()

Function takes in user objects and will create an array of
the display names associated with those users.