(feat) Fill in recipient and associated id for creating cache
After the user searches for their friend, they can click on the name
in the resulting list, and it will populate the field and
attach the associated uid on the properties object that is sent
over to Firebase.

Right now there is an issue in that the user could not click on
their friend's name, and then send, which would not send any
recipient over.