(fix) Display expiresAt time on live discovered cache
This fixes a possible race-condition where the cache might be marked
discovered before expiresAt was placed on the cache, meaning that the user
would not see the time till expiration until a refresh.

The fix is to only mark the cache as discovered AFTER the expiresAt property
has been set.

Fixes #119.