(feat) Immediately upload user-chosen image content
Calling addContent() from within the photo selection function keeps
us from uploading images that don't end up getting used if the user
decides to cancel out of the detail view without hitting the submit
button. Since there's already a confirmation within the image
selection and capture process, it's unnecessary to require the user
to confirm again. The current submit button is only for adding text-
only content, now.

Note: if a user enters text before selecting an image, that text will
still get uploaded as a caption with the image. We may wish to remove
that down the line.

- closes #196